requirements.txt
aiohappyeyeballs 
aiohttp 
aiosignal 
altair 
annotated-doc 
annotated-types 
anyio 
appnope 
asttokens 
async-timeout 
attrs 
beanie 
blinker 
branca 
Brotli 
cachetools 
certifi 
cffi 
charset-normalizer 
click 
comm 
cryptography 
debugpy 
decorator 
dnspython 
email-validator 
entrypoints 
exceptiongroup 
executing 
fastapi 
fastapi-cli 
folium 
frozenlist 
gitdb 
GitPython 
google-auth 
google-genai 
h11 
h2 
hpack 
httpcore 
httptools 
httpx 
httpx-sse 
hyperframe 
idna 
importlib_metadata 
ipykernel 
ipython 
ipython_pygments_lexers 
jedi 
Jinja2 
joblib==1.5.2
jsonschema 
jsonschema-specifications 
jupyter_client 
jupyter_core 
lazy-model 
mapclassify==2.10.0
markdown-it-py 
MarkupSafe 
matplotlib-inline 
mcp 
mdurl 
motor 
multidict 
nest_asyncio 
networkx==3.5
numpy 
packaging 
pandas 
parso 
pexpect 
pillow 
platformdirs 
prompt_toolkit 
propcache 
protobuf 
psutil 
ptyprocess 
pure_eval 
pyarrow==21.0.0
pyasn1 
pyasn1_modules 
pycparser 
pydantic 
pydantic-settings 
pydantic_core 
pydeck 
Pygments 
PyJWT 
pymongo 
pyOpenSSL 
PySocks 
python-dateutil 
python-dotenv 
python-multipart 
pytz 
pyu2f 
PyYAML 
pyzmq 
referencing 
requests 
rich 
rich-toolkit 
rpds-py 
rsa 
scikit-learn==1.7.2
scipy==1.16.3
shellingham 
six 
smmap 
sniffio 
sse-starlette 
stack_data 
starlette 
streamlit 
streamlit-folium 
tenacity 
threadpoolctl==3.6.0
toml 
tomli 
toolz 
tornado 
traitlets 
typer==0.20.0
typer-slim==0.20.0
typing-inspection 
typing_extensions 
tzdata 
urllib3 
uvicorn 
uvloop 
watchdog 
watchfiles 
wcwidth 
websockets 
xyzservices 
yarl 
zipp 
zstandard==0.25.0
